Conflicts:
	sys/kern/init_sysent.c
	sys/kern/syscalls.c
	sys/kern/syscalls.master
	sys/sys/syscall-hide.h
	sys/sys/syscall.h
	sys/sys/syscall.mk
	sys/sys/sysproto.h
	sys/sys/sysunion.h

+++ sys/vfs/ntfs/ntfs_subr.c
@@ -52,7 +52,6 @@
 #include <miscfs/specfs/specdev.h>
 #endif
 
-/* #define NTFS_DEBUG 1 */
 #include "ntfs.h"
 #include "ntfsmount.h"
 #include "ntfs_inode.h"
@@ -369,13 +368,13 @@ ntfs_ntlookup(struct ntfsmount *ntmp, ino_t ino, struct ntnode **ipp)
 {
 	struct ntnode  *ip;
 
-	dprintf(("ntfs_ntlookup: looking for ntnode %d\n", ino));
+	dprintf(("ntfs_ntlookup: looking for ntnode %ju\n", (uintmax_t)ino));
 
 	do {
 		if ((ip = ntfs_nthashlookup(ntmp->ntm_dev, ino)) != NULL) {
 			ntfs_ntget(ip);
-			dprintf(("ntfs_ntlookup: ntnode %d: %p, usecount: %d\n",
-				ino, ip, ip->i_usecount));
+			dprintf(("ntfs_ntlookup: ntnode %ju: %p, usecount: %d\n",
+				(uintmax_t)ino, ip, ip->i_usecount));
 			*ipp = ip;
 			return (0);
 		}
@@ -404,8 +403,8 @@ ntfs_ntlookup(struct ntfsmount *ntmp, ino_t ino, struct ntnode **ipp)
 
 	*ipp = ip;
 
-	dprintf(("ntfs_ntlookup: ntnode %d: %p, usecount: %d\n",
-		ino, ip, ip->i_usecount));
+	dprintf(("ntfs_ntlookup: ntnode %ju: %p, usecount: %d\n",
+		(uintmax_t)ino, ip, ip->i_usecount));
 
 	return (0);
 }
@@ -1160,7 +1159,7 @@ ntfs_ntreaddir(struct ntfsmount *ntmp, struct fnode *fp,
 
 	do {
 		dprintf(("ntfs_ntreaddir: scan: 0x%x, %d, %d, %d, %d\n",
-			 attrnum, (u_int32_t) blnum, cnum, num, aoff));
+			 attrnum, blnum, cnum, num, aoff));
 		rdsize = (attrnum == NTFS_A_INDXROOT) ? vap->va_datalen : blsize;
 		error = ntfs_readattr(ntmp, ip, attrnum, "$I30",
 				ntfs_cntob(blnum * cpbl), rdsize, rdbuf, NULL);
@@ -1215,7 +1214,7 @@ ntfs_ntreaddir(struct ntfsmount *ntmp, struct fnode *fp,
 			aoff = 0;
 			if (ntfs_cntob(blnum * cpbl) >= iavap->va_datalen)
 				break;
-			dprintf(("ntfs_ntreaddir: blnum: %d\n", (u_int32_t) blnum));
+			dprintf(("ntfs_ntreaddir: blnum: %d\n", blnum));
 		}
 	} while (iavap);
